These are code files to guide how to load the data
These files serve as a guide for loading data from four different datasets. Teams are free to modify them in any way that suits their needs. They are simply provided as examples to demonstrate how the process can be approached.These files serve as a guide for loading data from four different datasets. Teams are free to modify them in any way that suits their needs. They are simply provided as examples to demonstrate how the process can be approached.